backend/requirements.txt
Django>=5.2
djangorestframework>=3.14
requests>=2.31
